Introduce a docker compose file to run prometheus locally
Prometheus requires us to add volume, expose some ports, map those ports
to localhost, and communicate across docker images. This is all made a
lot easier when configured through docker-compose.

Quick note: Make sure the prometheus file is in a directory by itself,
and it is no higher in the file heirarchy than the docker-compose file.
This way you can map it as a volume. Also make sure the config file is
named _exactly_ prometheus.yml so that the server picks up the
configuration automatically.